syntax

+++ examples/simple_trainer.py
@@ -813,7 +813,8 @@ class Runner:
             #     with_eval3d=self.cfg.with_eval3d,
             # )
 
-            curr_cam = torch.from_numpy(np.linalg.inv(neuRISRunner.dataset.world_mats_np[image_ids.item()])).double().unsqueeze(0).to(device)
+            curr_cam = torch.from_numpy(np.linalg.inv(neuRISRunner.dataset.world_mats_np[image_ids.item()])).unsqueeze(0).to(device)
+            curr_cam = curr_cam.double()
             print("curr cam dtype ", curr_cam.dtype)
             tmp_renders, _, _ = self.render_depth_rasterize_splats(
                 #camtoworlds=camtoworlds,
